+++ bookiary-api/src/main/java/oz/bookiarybacked/domain/book/domain/model/Book.java
@@ -13,7 +13,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import oz.bookiarybacked.common.BaseTimeEntity;
+import oz.bookiarybacked.common.domain.model.BaseTimeEntity;
 import oz.bookiarybacked.domain.book.domain.dto.BookDto;
 import oz.bookiarybacked.domain.util.LocalDateUtils;
