+++ bookiary-api/src/main/java/oz/bookiarybacked/domain/note/domain/model/Note.java
@@ -11,7 +11,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import oz.bookiarybacked.common.BaseTimeEntity;
+import oz.bookiarybacked.common.domain.model.BaseTimeEntity;
 
 @Entity
 @Table(name = "note")
